+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Chore
 
+- BREAKING CHANGE: added sca inventory ([9dd49fd](https://github.com/cxpsemea/ts-inventory/commit/9dd49fd01e486765f988adf5e6aade649d70d84e) by jcostacx).
 - SCA Inventory Support ([73d0822](https://github.com/cxpsemea/ts-inventory/commit/73d0822a5d75aae43ce67de7208c48c82e0743fa) by jcostacx).
 - sast inventory ([b0b9f49](https://github.com/cxpsemea/ts-inventory/commit/b0b9f4915dff8cddcfd82df3987599dc7c2ead4e) by jcostacx).